==> 02-event-studies.R <==
library(eventstudies)
# https://r-forge.r-project.org/scm/viewvc.php/*checkout*/pkg/vignettes/eventstudies.Rnw?revision=339&root=eventstudies

# have a clean data script and source it here...
# for now, I'm picking up at the end of 01-randomness-tables.R
